extract chart functions to its own component file
